--- src/main/java/dev/ybw0014/guizhancraft/core/recipes/GuizhanCraftRecipeTypes.java
@@ -1,14 +1,0 @@
-package dev.ybw0014.guizhancraft.core.recipes;
-
-import io.github.thebusybiscuit.slimefun4.api.recipes.RecipeType;
-
-import lombok.experimental.UtilityClass;
-
-/**
- * This class holds all {@link RecipeType} in {@link dev.ybw0014.guizhancraft.implementation.GuizhanCraft}
- *
- * @author ybw0014
- */
-@UtilityClass
-public final class GuizhanCraftRecipeTypes {
-}
