--- src/main/java/dev/ybw0014/guizhancraft/implementation/setup/ItemSetup.java
@@ -1,46 +1,0 @@
-package dev.ybw0014.guizhancraft.implementation.setup;
-
-import org.bukkit.Material;
-import org.bukkit.inventory.ItemStack;
-
-import io.github.thebusybiscuit.slimefun4.api.items.SlimefunItem;
-import io.github.thebusybiscuit.slimefun4.api.recipes.RecipeType;
-import io.github.thebusybiscuit.slimefun4.implementation.SlimefunItems;
-
-import dev.ybw0014.guizhancraft.implementation.GuizhanCraft;
-import dev.ybw0014.guizhancraft.implementation.GuizhanCraftItems;
-import dev.ybw0014.guizhancraft.implementation.items.machine.BackpackUpgrader;
-import dev.ybw0014.guizhancraft.implementation.items.material.SimpleMaterial;
-
-import lombok.experimental.UtilityClass;
-
-/**
- * This class setup all {@link SlimefunItem} for GuizhanCraft
- *
- * @author ybw0014
- */
-@UtilityClass
-public final class ItemSetup {
-    public static void setup() {
-        GuizhanCraft plugin = GuizhanCraft.getInstance();
-
-        // region Materials
-        // endregion Materials
-
-        // region Machines
-
-        // endregion Machines
-    }
-
-    public static void setupElectricSpawners() {
-        GuizhanCraft plugin = GuizhanCraft.getInstance();
-
-        // region Materials
-
-        // endregion Materials
-
-        // region Machines
-
-        // endregion Machines
-    }
-}
