--- src/main/java/dev/ybw0014/guizhancraft/utils/HeadTextures.java
@@ -1,10 +1,0 @@
-package dev.ybw0014.guizhancraft.utils;
-
-import lombok.experimental.UtilityClass;
-
-@UtilityClass
-public final class HeadTextures {
-    public static final String MAIN_ITEM_GROUP = "15561bc0c5fa8e59f0799a24ce1bae7f43f1c4f330b3af8c663943a50623fa9";
-    public static final String MATERIAL_ITEM_GROUP = "c2deefe8579618bce23b541795eac6272f49f9790c0efe70744fb4f8ec3534c5";
-    public static final String MACHINE_ITEM_GROUP = "f7c243a9ced2fe4bd3cf616cf1a6456ae99bab98758babde6110bf3f3731eacb";
-}
