+++ src/main/java/net/guizhanss/guizhancraft/core/recipes/GuizhanCraftRecipeTypes.java
@@ -1,0 +1,15 @@
+package net.guizhanss.guizhancraft.core.recipes;
+
+import io.github.thebusybiscuit.slimefun4.api.recipes.RecipeType;
+
+import lombok.experimental.UtilityClass;
+import net.guizhanss.guizhancraft.implementation.GuizhanCraft;
+
+/**
+ * This class holds all {@link RecipeType} in {@link GuizhanCraft}
+ *
+ * @author ybw0014
+ */
+@UtilityClass
+public final class GuizhanCraftRecipeTypes {
+}
