+++ src/main/java/net/guizhanss/guizhancraft/implementation/items/material/SimpleMaterial.java
@@ -1,0 +1,19 @@
+package net.guizhanss.guizhancraft.implementation.items.material;
+
+import org.bukkit.inventory.ItemStack;
+
+import io.github.thebusybiscuit.slimefun4.api.items.ItemGroup;
+import io.github.thebusybiscuit.slimefun4.api.items.SlimefunItemStack;
+import io.github.thebusybiscuit.slimefun4.api.recipes.RecipeType;
+import io.github.thebusybiscuit.slimefun4.implementation.items.blocks.UnplaceableBlock;
+
+/**
+ * A simple unplaceable material
+ *
+ * @author ybw0014
+ */
+public class SimpleMaterial extends UnplaceableBlock {
+    public SimpleMaterial(ItemGroup itemGroup, SlimefunItemStack item, RecipeType recipeType, ItemStack[] recipe) {
+        super(itemGroup, item, recipeType, recipe);
+    }
+}
