+++ src/main/resources/lang/zh-CN.yml
@@ -8,28 +8,43 @@ items:
   electric_spawner_framework:
     name: "&e电力刷怪笼框架"
     lore:
-      - "&a好像差了点什么..."
+      - "&7好像差了点什么..."
   electric_spawner_assembler:
-    name: "&e&l增强&d古代祭坛"
+    name: "&d电力刷怪笼&a组装机"
     lore:
-      - "&a把祭坛和基座搓到一块"
-      - "&a到底还能不能用呢?"
+      - "&7可以用电力刷怪笼框架"
+      - "&7和已修复的刷怪笼"
+      - "&7组装电力刷怪笼"
+      - ""
+      - "&7可进行不消耗电力的手动合成"
 
 researches:
   electric_spawner_assembler: "电力刷怪笼组装机"
 
 gui:
-  preview: "&e预览"
   craft: "&a点击合成"
+  not-enough-energy: "&c电力不足"
+  no-input: "&c缺少输入"
+  no-output-space: "&c输出空间不足"
+  success: "&a成功!"
+  electric_spawner_assembler:
+    info:
+      name: "&e说明"
+      lore:
+        - "&b在左侧放入&e电力刷怪笼框架"
+        - "&b在右侧放入&e已修复的刷怪笼"
+        - ""
+        - "&b点击下方的按钮可手动合成一次"
+        - "&b(无需电力)"
 
 messages:
-  no_permission: "&4你没有权限!"
-  no_console: "&4只有玩家才能执行该指令!"
+  no-permission: "&4你没有权限!"
+  no-console: "&4只有玩家才能执行该指令!"
   usage: "&e用法: &a{0}"
-  player_not_found: "&c找不到指定玩家 {0}!"
-  invalid_recipe: "&c无效的配方!"
-  input_all: "&c请在所有输入栏放入物品!"
-  output_no_space: "&c输出栏已满!"
+  player-not-found: "&c找不到指定玩家 {0}!"
+  invalid-recipe: "&c无效的配方!"
+  no-input: "&c请在所有输入栏放入物品!"
+  no-output-space: "&c输出空间不足!"
   crafted: "&a成功合成&e{0}&a!"
 
 console:
@@ -38,11 +53,12 @@ console:
   loading_commands: "加载指令中..."
   loading_researches: "研究已启用, 加载研究中..."
   loading_integrations: "加载关联插件中..."
-  loading_integration: "已关联插件 {0} v{1}"
+  loading_integration: "正在关联插件 {0}"
+  loaded_integration: "已关联插件 {0} v{1}"
 
 commands:
   sfspawner:
-    invalid_entity_type: "&c无效的生物类型: {0}"
+    invalid-entity-type: "&c无效的生物类型: {0}"
     sent: "&b{0} &a已获得 &b{1}"
     received: "&a你收到了来自 &b{0} &a的 &b{1}"
     broken:
